requirements.txt
flask==2.0.1
requests==2.28.1
python-dotenv==0.20.0
apscheduler==3.9.1
flask-cors==3.0.10
gunicorn==20.1.0